RINGUS-10 refactor: 회원가입 반환 DTO 필드 및 컨트롤러 반환 엔티티 수정
