--- src/main/java/es/princip/ringus/application/auth/dto/SignUpResponse.java
@@ -1,8 +1,0 @@
-package es.princip.ringus.application.auth.dto;
-
-public record SignUpResponse(
-        int status,
-        String message,
-        Object data
-) {
-}
